renamed is_sort.py -> is_sorted.py and add tests.
