+++ linkedlist/is_sorted.py
@@ -1,0 +1,56 @@
+"""
+Given a linked list, is_sort function returns true if the list is in sorted
+(increasing) order and return false otherwise. An empty list is considered
+to be sorted.
+
+For example:
+Null :List is sorted
+1 2 3 4 :List is sorted
+1 2 -1 3 :List is not sorted
+"""
+import unittest
+
+
+class Node:
+    def __init__(self, x):
+        self.val = x
+        self.next = None
+
+
+def is_sorted(head):
+    if not head:
+        return True
+    current = head
+    while current.next:
+        if current.val > current.next.val:
+            return False
+        current = current.next
+    return True
+
+
+class TestSuite(unittest.TestCase):
+
+    def test_is_sorted(self):
+
+        head = Node(-2)
+        head.next = Node(2)
+        head.next.next = Node(2)
+        head.next.next.next = Node(4)
+        head.next.next.next.next = Node(9)
+
+        # head -> -2 -> 2 -> 2 -> 4 -> 9
+        self.assertTrue(is_sorted(head))
+
+        head = Node(1)
+        head.next = Node(2)
+        head.next.next = Node(8)
+        head.next.next.next = Node(4)
+        head.next.next.next.next = Node(6)
+
+        # head -> 1 -> 2 -> 8 -> 4 -> 6
+        self.assertFalse(is_sorted(head))
+
+
+if __name__ == "__main__":
+
+    unittest.main()
